rename heartbeats, use healthchecks instead
With the introduction of receptor nodes, we want to ensure the
overall health of a node, instead of just checking to make sure the
node is responsive and has capacity.

A 'healthy' node is one that has all the requirements for running a job
(podman, ansible-runner) in addition to having cpu/memory resources
and a receptor configuration that lists the correct worktype.